Update to ChakraCore v1.11.8

+++ tests/Test_ChakraCore.pas
@@ -168,7 +168,7 @@ procedure TChakraCoreUtilsScripting.TestVersion;
 begin
   CheckEquals(Integer(1), CHAKRA_CORE_MAJOR_VERSION, 'major version number');
   CheckEquals(Integer(11), CHAKRA_CORE_MINOR_VERSION, 'minor version number');
-  CheckEquals(Integer(7), CHAKRA_CORE_PATCH_VERSION, 'patch version number');
+  CheckEquals(Integer(8), CHAKRA_CORE_PATCH_VERSION, 'patch version number');
 end;
 
 procedure TChakraCoreUtilsScripting.TestUndefined;
